requirements.txt
langgraph
langchain
langchain-deepseek
fastapi
uvicorn[standard]
python-dotenv
watchdog
rank_bm25
python-frontmatter
sqlalchemy
pydantic
beautifulsoup4